requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.1.2
click==8.2.1
cloud-sql-python-connector==1.18.5
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dnspython==2.8.0
dtw-python==1.7.2
et_xmlfile==2.0.0
fonttools==4.58.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.41.1
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==1.1.0
kiwisolver==1.4.8
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.7.0
mysql-connector-python==9.4.0
narwhals==1.42.0
numpy==2.2.5
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==24.2
pandas==2.2.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
plotly==6.2.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
PyMySQL==1.1.2
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
scipy==1.16.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.22.0
